% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_vec.R
\name{fmt_count_prop}
\alias{fmt_count_prop}
\title{Inline format of counts with proportion.}
\usage{
fmt_count_prop(x, ...)
}
\arguments{
\item{x}{a dataframe object.}

\item{...}{a condition for filtering x.}
}
\value{
inline counts and proportions for condition.
}
\description{
Inline format of counts with proportion.
}
