% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugz.utils.R
\name{fmt_n}
\alias{fmt_n}
\title{Format counts using K and M symbols.}
\usage{
fmt_n(n)
}
\arguments{
\item{n}{a numeric vector.}
}
\value{
formatted values using M for millions and K for thousands
}
\description{
Format counts using K and M symbols.
}
