% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_vec.R
\name{get_id_vec}
\alias{get_id_vec}
\title{get a vector of id from df given condition.}
\usage{
get_id_vec(df, ..., id_var = "pid")
}
\arguments{
\item{df}{a dataframe.}

\item{...}{a condition to filter a dataframe.}

\item{id_var}{specifiy id variable, default = pid.}
}
\value{
a character vector of unique id_var after filtering for condition.
}
\description{
get a vector of id from df given condition.
}
\examples{

dat <- data.frame( "pid" = c("xx-1", "xx-2", "xx-3"), "age" = c(12, 20, 39) )

get_id_vec(dat, age < 30)
}
