% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugz.utils.R
\name{get_interval}
\alias{get_interval}
\title{Time Interval between two dates.}
\usage{
get_interval(x1, x2, unit)
}
\arguments{
\item{x1}{first date vector.}

\item{x2}{second date vector.}

\item{unit}{units of time for interval. One of "days", "weeks", "months", "years".}
}
\value{
date interval using {lubridate}.
}
\description{
Time Interval between two dates.
}
