% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugz.utils.R
\name{get_msf_onedrive}
\alias{get_msf_onedrive}
\title{Construct path to MSF onedrive.}
\usage{
get_msf_onedrive(project = NULL, path = NULL)
}
\arguments{
\item{project}{specify the projects from MSF one drive.}

\item{path}{lower level path to fetch from.}
}
\description{
Construct path to MSF onedrive.
}
