% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugz.utils.R
\name{write_multi_xlsx}
\alias{write_multi_xlsx}
\title{Title}
\usage{
write_multi_xlsx(
  df,
  df_name,
  extra_sheet,
  extra_sheet_name,
  file_path,
  file_name
)
}
\arguments{
\item{df}{a df to save to .xlsx}

\item{df_name}{a name for the main sheet}

\item{extra_sheet}{a df to be added as a second sheet}

\item{extra_sheet_name}{a name for the second sheet}

\item{file_path}{path to save the file}

\item{file_name}{file name using .xlsx
ss}
}
\description{
Title
}
